Can now be deployed as WAR using boot dashboard
See
https://issuetracker.springsource.com/browse/STS-4411?focusedCommentId=112813&page=com.atlassian.jira.plugin.system.issuetabpanels:comment-tabpanel#comment-112813
for a comment about how this can work although this is not a good
permanent solution because it relies only on circumstantial setup that
could change in the future.

Home controller is to redirect to music.jsp.  I use jsp to process the
query parameter using the container servlet stack instead of parsing it
in javascript.

NotationMsterModel.js: